\name{extended.figures}
\alias{extended.figures}
\title{Generate figures from An Extension of Wilkinson's Algorithm for Position Tick Labels on Axes}
\usage{
  extended.figures(samples = 100)
}
\arguments{
  \item{samples}{number of samples to use (in the paper we
  used 10000, but that takes awhile to run).}
}
\value{
  produces plots as a side effect
}
\description{
  Generates Figures 2 and 3 from our paper.
}
\author{
  Justin Talbot \email{justintalbot@gmail.com}
}
\references{
  Talbot, J., Lin, S., Hanrahan, P. (2010) An Extension of
  Wilkinson's Algorithm for Positioning Tick Labels on
  Axes, InfoVis 2010.
}

