\name{gnuplot}
\alias{gnuplot}
\title{gnuplot's labeling algorithm}
\usage{
  gnuplot(dmin, dmax, m)
}
\arguments{
  \item{dmin}{minimum of the data range}

  \item{dmax}{maximum of the data range}

  \item{m}{number of axis labels}
}
\value{
  vector of axis label locations
}
\description{
  gnuplot's labeling algorithm
}
\author{
  Justin Talbot \email{justintalbot@gmail.com}
}
\references{
  \url{http://www.gnuplot.info/}
}

