\name{heckbert}
\alias{heckbert}
\title{Heckbert's labeling algorithm}
\usage{
  heckbert(dmin, dmax, m)
}
\arguments{
  \item{dmin}{minimum of the data range}

  \item{dmax}{maximum of the data range}

  \item{m}{number of axis labels}
}
\value{
  vector of axis label locations
}
\description{
  Heckbert's labeling algorithm
}
\author{
  Justin Talbot \email{justintalbot@gmail.com}
}
\references{
  Heckbert, P. S. (1990) Nice numbers for graph labels,
  Graphics Gems I, Academic Press Professional, Inc.
}

