\name{thayer}
\alias{thayer}
\title{Thayer and Storer's labeling algorithm}
\usage{
  thayer(dmin, dmax, m)
}
\arguments{
  \item{dmin}{minimum of the data range}

  \item{dmax}{maximum of the data range}

  \item{m}{number of axis labels}
}
\value{
  vector of axis label locations
}
\description{
  Thayer and Storer's labeling algorithm
}
\author{
  Justin Talbot \email{justintalbot@gmail.com}
}
\references{
  Thayer, R. P. and Storer, R. F. (1969) AS 21. Scale
  Selection for Computer Plots, Journal of the Royal
  Statistical Society. Series C., pp. 206-208.
}

